+++ R/layers-core.R
@@ -259,6 +259,9 @@ layer_lambda <- function(object, f, output_shape = NULL, mask = NULL, arguments 
   
   if (backend()$backend() %in% c("theano", "cntk"))
     args$output_shape <- as_integer_tuple(output_shape, force_tuple = TRUE)
+  else if(!is.null(output_shape))
+    args$output_shape <- normalize_shape(output_shape)
+    
   
   create_layer(keras$layers$Lambda, object, args)
   
@@ -340,15 +343,19 @@ layer_masking <- function(object, mask_value = 0.0, input_shape = NULL,
 #'
 #' @inheritParams layer_activation
 #'
-#' @param data_format A string, one of `channels_last` (default) or
-#'   `channels_first`. The ordering of the dimensions in the inputs.
-#'   `channels_last` corresponds to inputs with shape `(batch, ..., channels)`
-#'   while `channels_first` corresponds to inputs with shape `(batch, channels, ...)`.
+#' @param data_format A string. one of `channels_last` (default) or
+#'   `channels_first`. The ordering of the dimensions in the inputs. The purpose
+#'   of this argument is to preserve weight ordering when switching a model from
+#'   one data format to another. `channels_last` corresponds to inputs with
+#'   shape `(batch, ..., channels)` while `channels_first` corresponds to inputs
+#'   with shape `(batch, channels, ...)`. It defaults to the `image_data_format`
+#'   value found in your Keras config file at `~/.keras/keras.json`. If you
+#'   never set it, then it will be "channels_last".
 #'
 #' @family core layers
 #'
 #' @export
-layer_flatten <- function(object, data_format = "channels_last", input_shape = NULL, dtype = NULL, 
+layer_flatten <- function(object, data_format = NULL, input_shape = NULL, dtype = NULL, 
                           name = NULL, trainable = NULL, weights = NULL) {
   
   args <- list(
@@ -359,8 +366,13 @@ layer_flatten <- function(object, data_format = "channels_last", input_shape = N
     weights = weights
   )
   
-  if (keras_version() >= "2.1.6")
+  if (keras_version() >= "2.2.0") {
     args$data_format <- data_format
+  } else if (keras_version() >= "2.1.6") {
+    if (is.null(data_format))
+      data_format <- "channels_last"
+    args$data_format <- data_format
+  }
   
   create_layer(keras$layers$Flatten, object, args)
   
@@ -393,11 +405,11 @@ as_nullable_integer <- function(x) {
     as.integer(x)
 }
 
-as_nullable_array <- function(x) {
-  if (is.null(x))
-    x
-  else
-    as.array(x)
+as_layer_index <- function(x) {
+  x <- as_nullable_integer(x)
+  if (!is.null(x))
+    x <- x - 1L
+  x
 }
 
 # Helper function to normalize paths